Use intptr_t for asm function parameters
int can be a problem for arguments passed on the stack when the compiler
decides to not touch the high 32 bits of the stack element. This has
happened to VapourSynth in 64 bit Windows and maybe to x264 in 64 bit
Linux when compiled with clang.

Also call certain functions from Interpolation.asm with the correct
number of parameters. The stride was passed twice and it was working
only by coincidence.