MDegrain - Simplify SSE2 code using loops.
I tested this with a 4k, 2k, and 540p video. I could find no discernable
difference in performance between the two, and it takes up a lot less
code.

Results:
Degrain6 SSE2 WITHOUT loops on 8-bit, 1000 frames
540p = 77.5433
1080p = 50.5460
4k = 12.0173

Degrain6 SSE2 WITH loops on 8-bit, 1000 frames
540p = 79.3693
1080p = 49.8574
4k = 11.9895

This is with GCC 12.2.0 on Arch Linux.